[Silver III] Title: 주유소, Time: 52 ms, Memory: 2800 KB -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.cc
@@ -26,7 +26,7 @@ int main() {
 
 	int curPrice = price[1];
 	int curIdx = 2;
-	int total = 0;
+	long long total = 0;
 
 	while (curIdx <= N) {
 		if (curPrice >= price[curIdx]) {
